2/6/2021 Eric T
I made these changes weeks ago, I forgot what needs to be added. But the
most important thing right now, is to not have to permantly load
characters into the uesr list. What needs to be done is the user file
only needs to be loaded and saved to whenever an action invovles that
user.